prev debug

+++ server/irods_server.py
@@ -271,7 +271,7 @@ def retreive_matching_file_list(patient_dir, file_age, search_terms):
     all_files = subprocess.run(cmd, shell=True, stdout=subprocess.PIPE).stdout.decode('utf-8').splitlines()
     for file in all_files:
         cmd = f"imeta ls -d '{file}' date_create | awk '/value/ {{print $2}}'"
-        output = long(subprocess.run(cmd, shell=True, stdout=subprocess.PIPE).stdout.decode('utf-8'))
+        output = int(subprocess.run(cmd, shell=True, stdout=subprocess.PIPE).stdout.decode('utf-8'))
         if output >= age:
             file_matches.append(file)
